pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-recognition"
version = "0.1.0"
description = "Sistema de Reconhecimento Facial usando MLP"
authors = [
    {name = "Ana Beatriz, Ana Paula, Bernardo Rosa, Livia Rainho"}
]
requires-python = ">=3.11"
dependencies = [
    "Pillow>=8.0.0",
    "opencv-python-headless>=4.5.0",
    "opencv-contrib-python>=4.5.0",
    "numpy>=1.24.0,<2.0.0",
    "insightface>=0.7.0",
    "onnxruntime-gpu>=1.8.0",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
cuda = [
    "torch @ https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp311-cp311-win_amd64.whl",
    "torchvision @ https://download.pytorch.org/whl/cu118/torchvision-0.15.2%2Bcu118-cp311-cp311-win_amd64.whl"
]

[project.scripts]
train-face-model = "train_model:main"
predict-face = "predictor:main"
detect-faces = "face_detector:main"
realtime-recognition = "real_time_recognition:main"
download-models = "download_models:main"

[tool.setuptools]
py-modules = ["train_model", "predictor", "face_detector", "real_time_recognition", "download_models"]
